Create a new OAuth resource, refactor OpenEdxApiClient to inherit from it
