+++ src/ol_orchestrate/definitions/learning_resource/extract_api_data.py
@@ -10,7 +10,7 @@ from ol_orchestrate.io_managers.filepath import S3FileObjectIOManager
 from ol_orchestrate.lib.constants import DAGSTER_ENV, VAULT_ADDRESS
 from ol_orchestrate.lib.dagster_helpers import default_io_manager
 from ol_orchestrate.lib.utils import authenticate_vault, s3_uploads_bucket
-from ol_orchestrate.resources.openedx import OpenEdxApiClientFactory
+from ol_orchestrate.resources.oauth import OAuthApiClientFactory
 
 vault = authenticate_vault(DAGSTER_ENV, VAULT_ADDRESS)
 
@@ -30,9 +30,7 @@ extract_api_data = Definitions(
         ),
         "vault": vault,
         "s3": S3Resource(),
-        "sloan_api": OpenEdxApiClientFactory(
-            deployment="sloan-executive-education", vault=vault
-        ),
+        "sloan_api": OAuthApiClientFactory(deployment="sloan", vault=vault),
     },
     assets=[sloan_course_metadata],
     schedules=[extract_api_daily_schedule],
