+++ src/ol_orchestrate/resources/openedx.py
@@ -1,123 +1,26 @@
-import time
 from collections.abc import Generator
 from contextlib import contextmanager
-from datetime import UTC, datetime, timedelta
-from typing import Any, Optional, Self
+from typing import Optional, Self
 from urllib.parse import parse_qs
 
-import httpx
 from dagster import ConfigurableResource, InitResourceContext, ResourceDependency
-from pydantic import Field, PrivateAttr, ValidationError, validator
+from pydantic import Field, PrivateAttr
 
+from ol_orchestrate.resources.oauth import OAuthApiClient
 from ol_orchestrate.resources.secrets.vault import Vault
 
 TOO_MANY_REQUESTS = 429
 
 
-class OpenEdxApiClient(ConfigurableResource):
-    client_id: str = Field(description="OAUTH2 client ID for Open edX installation")
-    client_secret: str = Field(
-        description="OAUTH2 client secret for Open edX installation"
-    )
-    lms_url: str = Field(
-        description="Base URL of edX instance LMS being queried, including protocol. "
-        "e.g. https://lms.mitx.mit.edu"
-    )
+class OpenEdxApiClient(OAuthApiClient):
     studio_url: Optional[str] = Field(
         default=None,
         description="Base URL of edx instance Studio being queried, including protocol."
         " e.g. https://studio.mitx.mit.edu",
     )
-    token_type: str = Field(
-        default="JWT",
-        description="Token type to generate for use with authenticated requests",
-    )
-    token_url: str = Field(
-        description="URL to request token. e.g. https://lms.mitx.mit.edu/oauth2/access_token",
-    )
-    http_timeout: int = Field(
-        default=60,
-        description=(
-            "Time (in seconds) to allow for requests to complete before timing out."
-        ),
-    )
-    _access_token: Optional[str] = PrivateAttr(default=None)
-    _access_token_expires: Optional[datetime] = PrivateAttr(default=None)
-    _http_client: httpx.Client = PrivateAttr(default=None)
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self._initialize_client()
-
-    @validator("token_type")
-    def validate_token_type(cls, token_type):  # noqa: N805
-        if token_type.lower() not in ["jwt", "bearer"]:
-            raise ValidationError
-        return token_type
-
-    def _initialize_client(self) -> None:
-        if self._http_client is not None:
-            return
-        timeout = httpx.Timeout(self.http_timeout, connect=10)
-        self._http_client = httpx.Client(timeout=timeout)
-
-    def _fetch_access_token(self) -> Optional[str]:
-        now = datetime.now(tz=UTC)
-        if self._access_token is None or (self._access_token_expires or now) <= now:
-            payload = {
-                "grant_type": "client_credentials",
-                "client_id": self.client_id,
-                "client_secret": self.client_secret,
-                "token_type": self.token_type,
-            }
-            response = self._http_client.post(self.token_url, data=payload)
-            response.raise_for_status()
-            self._access_token = response.json()["access_token"]
-            self._access_token_expires = now + timedelta(
-                seconds=response.json()["expires_in"]
-            )
-        return self._access_token
-
-    @property
-    def _username(self) -> str:
-        response = self._http_client.get(
-            f"{self.lms_url}/api/user/v1/me",
-            headers={"Authorization": f"JWT {self._fetch_access_token()}"},
-        )
-        response.raise_for_status()
-        return response.json()["username"]
-
-    def _fetch_with_auth(
-        self,
-        request_url: str,
-        page_size: int = 100,
-        extra_params: dict[str, Any] | None = None,
-    ) -> dict[Any, Any]:
-        if self.token_url == f"{self.lms_url}/oauth2/access_token":
-            request_params = {"username": self._username, "page_size": page_size}
-        else:
-            request_params = {}
-
-        request_params.update(**(extra_params or {}))
-
-        response = self._http_client.get(
-            request_url,
-            headers={"Authorization": f"JWT {self._fetch_access_token()}"},
-            params=httpx.QueryParams(**request_params),
-        )
-
-        try:
-            response.raise_for_status()
-        except httpx.HTTPStatusError as error_response:
-            if error_response.response.status_code == TOO_MANY_REQUESTS:
-                retry_after = error_response.response.headers.get("Retry-After", 60)
-                delay = int(retry_after) if retry_after.isdigit() else 60
-                time.sleep(delay)
-                return self._fetch_with_auth(
-                    request_url, page_size=page_size, extra_params=extra_params
-                )
-            raise
-        return response.json()
 
     def get_edx_course_ids(
         self, page_size: int = 100
@@ -130,7 +33,7 @@ class OpenEdxApiClient(ConfigurableResource):
         :yield: A generator for walking the paginated list of courses returned from
             the API
         """
-        request_url = f"{self.lms_url}/api/courses/v1/courses/"
+        request_url = f"{self.base_url}/api/courses/v1/courses/"
         response_data = self._fetch_with_auth(request_url, page_size=page_size)
         course_data = response_data["results"]
         next_page = response_data["pagination"].get("next")
@@ -175,12 +78,12 @@ class OpenEdxApiClient(ConfigurableResource):
         :returns: JSON document representing the hierarchical structure of the target
                   course.
         """
-        request_url = f"{self.lms_url}/api/course-structure/v0/{course_id}/"
+        request_url = f"{self.base_url}/api/course-structure/v0/{course_id}/"
         return self._fetch_with_auth(request_url)
 
     def get_course_outline(self, course_id: str):
         request_url = (
-            f"{self.lms_url}/api/learning_sequences/v1/course_outline/{course_id}"
+            f"{self.base_url}/api/learning_sequences/v1/course_outline/{course_id}"
         )
         return self._fetch_with_auth(request_url)
 
@@ -226,24 +129,6 @@ class OpenEdxApiClient(ConfigurableResource):
             next_page = response_data["next"]
             yield response_data["results"]
 
-    def get_sloan_courses(self):
-        """
-        Retrieve the course data from their API as JSON
-
-        returns: JSON document representing an array of course objects
-        """
-        course_url = "https://mit-unified-portal-prod-78eeds.43d8q2.usa-e2.cloudhub.io/api/courses"
-        return self._fetch_with_auth(course_url)
-
-    def get_sloan_course_offerings(self):
-        """
-        Retrieve the course offerings data from their API as JSON
-
-        returns: JSON document representing an array of course offering objects
-        """
-        course_offering_url = "https://mit-unified-portal-prod-78eeds.43d8q2.usa-e2.cloudhub.io/api/course-offerings"
-        return self._fetch_with_auth(course_offering_url)
-
 
 class OpenEdxApiClientFactory(ConfigurableResource):
     deployment: str = Field(
@@ -262,7 +147,7 @@ class OpenEdxApiClientFactory(ConfigurableResource):
         self._client = OpenEdxApiClient(
             client_id=client_secrets["id"],
             client_secret=client_secrets["secret"],
-            lms_url=client_secrets["url"],
+            base_url=client_secrets["url"],
             studio_url=client_secrets["studio_url"],
             token_url=client_secrets.get(
                 "token_url", f"{client_secrets['url']}/oauth2/access_token"
